requirements.txt
tensorflow==2.4.0
streamlit==0.67.1
numpy==1.18.5
opencv-python
pillow==7.2.0
